requirements.txt
alabaster==0.7.11
astroid==2.2.5
Babel==2.6.0
certifi==2018.8.24
chardet==3.0.4
Django==2.2.2
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
docutils==0.14
idna==2.7
imagesize==1.0.0
isort==4.3.20
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.0
mccabe==0.6.1
packaging==17.1
pbr==4.2.0
Pillow==6.0.0
pipenv==2018.7.1
psycopg2==2.8.2
Pygments==2.2.0
PyJWT==1.7.1
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
pyparsing==2.2.0
pytz==2018.5
requests==2.20.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.8
sphinx-rtd-theme==0.4.1
sphinxcontrib-websupport==1.1.0
sqlparse==0.3.0
stevedore==1.29.0
typed-ast==1.3.5
urllib3==1.24.2
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
wrapt==1.11.1
django-cors-headers==3.0.2
ShopifyAPI==5.1.0
boto3==1.9.223
